requirements.txt
tensorflow>=2.9.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
pandas>=1.3.0
seaborn>=0.11.0
pillow>=9.0.0
pillow-heif>=0.10.0
pyheif>=0.7.0